% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_result.R
\name{cleanup_result}
\alias{cleanup_result}
\title{cleanup_result}
\usage{
cleanup_result(res_all)
}
\arguments{
\item{res_all}{a list from function 'run_bit'}
}
\value{
A list with 2 units. One is the rowindex of df, which shows the units that are matched. Another is a dataframe contains the size and the ate from each iteration of FLAME.
}
\description{
clean up the results of function 'run_bit'
}
\details{
This is some descriptio of this function.

Use the result from 'run_bit'. Then use function 'recover_covs' to get ATE.
}
