% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame.R
\name{flame}
\alias{flame}
\title{flame}
\usage{
flame(df, holdout, covs, covs_max_list, tradeoff_param = 0.1, printcov = FALSE)
}
\arguments{
\item{df}{df is a dataframe of dataset including covs, outcome and treated.}

\item{holdout}{holdout is a dataframe including data with the same distribution as df.}

\item{covs}{A subset of colnames of df, which shows the covs this match will use.}

\item{covs_max_list}{A a numric vector which shows the base system of cov.}

\item{tradeoff_param}{The parameter that trade off BF and PE.}

\item{printcov}{If TRUE, the name of used covs in each iteration will be printed.}
}
\value{
the final estimate ATE value
}
\description{
match units according to the algrithm of FLAME
}
\details{
This is some descriptio of this function.

First calculate the bit-vector value. Then count the number of times they appear. Judge whether they are equal.
}
