% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CATE_bit.R
\name{get_CATE_bit}
\alias{get_CATE_bit}
\title{get_CATE_bit}
\usage{
get_CATE_bit(df, match_indicator, index)
}
\arguments{
\item{df}{df is a dataframe of dataset including covs, outcome and treated.}

\item{match_indicator}{the index of units that are matched}

\item{index}{the bit-vector value}
}
\value{
a dataframe including size and mean-value
}
\description{
group units by bit-vector value and 'treated'
}
\details{
This is some descriptio of this function.

group units by bit-vector value and 'treated' so that ATE will be easier to get
}
