% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimate_vectors.R
\name{get_estimate_vectors}
\alias{get_estimate_vectors}
\title{get_estimate_vectors}
\usage{
get_estimate_vectors(result)
}
\arguments{
\item{result}{the result of funciont 'cleanup_result'}
}
\value{
Nothing will be returned.
}
\description{
print the ATE
}
\details{
This is some descriptio of this function.

Input the result from function 'cleanup_result', then print the ATE.
}
