% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match}
\alias{match}
\title{match}
\usage{
match(df, covs, covs_max_list)
}
\arguments{
\item{df}{df is a dataframe of dataset including covs, outcome and treated.}

\item{covs}{A subset of colnames of df, which shows the covs this match will use.}

\item{covs_max_list}{A a numric vector which shows the base system of cov.}
}
\value{
A list with 2 units. One is the rowindex of df, which shows the units that are matched. Another is the bit-vector value of these units.
}
\description{
match units according to the algrithm of FLAME
}
\details{
This is some descriptio of this function.

First calculate the bit-vector value. Then count the number of times they appear. Judge whether they are equal.
}
