% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_quality.R
\name{match_quality}
\alias{match_quality}
\title{match_quality}
\usage{
match_quality(df, holdout, covs_subset, match_indicator, tradeoff = 0.1)
}
\arguments{
\item{df}{df is a dataframe of dataset including covs, outcome and treated.}

\item{holdout}{holdout is a dataframe including data the same distribution with df.}

\item{covs_subset}{The names of covs that will be used in this match.}

\item{match_indicator}{the index of units that are matched}

\item{tradeoff}{The parameter that trade off BF and PE.}
}
\value{
A list with 3 units. First is the score. Second is the balance factor. Third is the prediction error.
}
\description{
compute the match quality score of one match
}
\details{
This is some descriptio of this function.

Compute the PE and BF, then return the score.
}
