% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bit.R
\name{run_bit}
\alias{run_bit}
\title{run_bit}
\usage{
run_bit(
  df,
  holdout,
  covs,
  covs_max_list,
  tradeoff_param = 0.1,
  printcov = TRUE
)
}
\arguments{
\item{df}{df is a dataframe of dataset including covs, outcome and treated.}

\item{holdout}{holdout is a dataframe including data with the same distribution as df.}

\item{covs}{A subset of colnames of df, which shows the covs this match will use.}

\item{covs_max_list}{A a numric vector which shows the base system of cov.}

\item{tradeoff_param}{The parameter that trade off BF and PE.}

\item{printcov}{If TRUE, the name of used covs in each iteration will be printed.}
}
\value{
A list with 2 units. One is the matching result in each iteration. Another is match quality score in each iteration.
}
\description{
run the FLAME algrithm
}
\details{
This is some descriptio of this function.

Drop one cov in each iteration until the prediction error is too large. Use the remaining covs to get the estiamte of ATE.
}
