\name{StatComp21035-package}
\alias{StatComp21035-package}
\alias{StatComp21035}
\docType{package}
\title{
  Fast Large-scale Almost Exactly Matching
}
\description{
  This package uses bit-vector FLAME methods to drop irrelevant covarates to decrease the bias of ATE estimation.
}
\details{
  In each iteration, the covarate with the highest score will be dropped. The score is affected by balance factor and prediction error. When the PE is too large, the iteration process stop.
}
\author{
Ma Haotian

Maintainer: Ma Haotian <mht1130@mail.ustc.edu.cn>
}
\keyword{ package }

